feat: support for plain static vite dist folder
