+++ examples/demo/tests/04-isr/globalSetup.ts
@@ -22,6 +22,7 @@ export const setup = _setup(path.basename(__dirname), {
         symlink: 'api/page',
       },
     },
+    distContainsOnlyStatic: false,
   },
 });
